Add files via upload
This data is a sample of the Nebraska COVID data provided by the The Covid Tracking project at the Atlantic for use in our class to learn command line file manipulation.  The COVID Tracking Project at The Atlantic’s data and website content is published under a Creative Commons CC BY 4.0 license, which requires users to attribute the source and license type (CC BY 4.0) when sharing our data or website content. See https://covidtracking.com/about-data/license for more information. This file is intended for learning command line processes in BMI 8540 and has been edited for that purpose.